requirements.txt

httpx>=0.27
pandas>=2.2
numpy>=1.26
pydantic>=2.9
python-dotenv>=1.0
websockets>=12.0
pyyaml>=6.0
